refactor: change Dashboard to Home
